# docs/swagger.yaml
definitions:
  models.Article:
    properties:
      a:
        $ref: '#/definitions/models.Person'
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - a
    - body
    - title
    type: object
  models.CreateArticleModel:
    properties:
      a:
        $ref: '#/definitions/models.Person'
      body:
        type: string
      title:
        type: string
    required:
    - a
    - body
    - title
    type: object
  models.JSONErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.JSONResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  models.Person:
    properties:
      firstname:
        example: John
        maxLength: 50
        minLength: 2
        type: string
      lastname:
        example: Doe
        maxLength: 50
        minLength: 2
        type: string
    required:
    - firstname
    - lastname
    type: object
info:
  contact:
    email: john.doe@swagger.io
    name: API Article
    url: https://john.doe.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /v2/article:
    get:
      consumes:
      - application/json
      description: get articles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Article'
                  type: array
              type: object
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: create a new article
      parameters:
      - description: article body
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.CreateArticleModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Article'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: Create article
      tags:
      - articles
  /v2/article/{id}:
    get:
      consumes:
      - application/json
      description: get an article by id
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Article'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONErrorResponse'
      summary: get article by id
      tags:
      - articles
swagger: "2.0"